Remove all priorities with `_`
instead of `=0`. The new way is more visual and =0 seems very
artificial as there is no =1 and a shift of an unprioritized task, stays
unprioritized which is strange if you think of the priorities as scalar
numbers.

(FYI: Though if priorities are vectors it would match =0 mathematically:
unprioritized is (0,0), A = (1, 0) and a shift by 1 is the rotation by
pi/13 . But I regress...)